# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get users list",
                "operationId": "566aed8efb436ca506e0ef51a032ecd9",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "User id",
                                                        "type": "string"
                                                    },
                                                    "fio": {
                                                        "description": "User fio",
                                                        "type": "string"
                                                    },
                                                    "birth_date": {
                                                        "description": "User birth date",
                                                        "type": "string"
                                                    },
                                                    "group_id": {
                                                        "description": "User group id",
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "description": "Creating date",
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "updated_at": {
                                                        "description": "Updating date",
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "address": {
                                                        "description": "User address",
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "description": "User email",
                                                        "type": "string"
                                                    },
                                                    "password": {
                                                        "description": "User password",
                                                        "type": "string"
                                                    },
                                                    "role": {
                                                        "description": "User role",
                                                        "type": "string"
                                                    },
                                                    "avatar": {
                                                        "description": "User avatar path",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Store user",
                "operationId": "be551c1d694a01c164966f58bfa77013",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fio": {
                                        "description": "User fio",
                                        "type": "string"
                                    },
                                    "birth_date": {
                                        "description": "User birth day",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "group_id": {
                                        "description": "User group id",
                                        "type": "integer"
                                    },
                                    "role": {
                                        "description": "User role",
                                        "type": "integer"
                                    },
                                    "address": {
                                        "description": "User address",
                                        "type": "json"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/users/{user}/export": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Export pdf file",
                "operationId": "1a8ba102b58ab63dd5ab560b4fa40fb5",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "description": "Url to downloading pdf",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/restore": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Restore user",
                "operationId": "13b5c02c3527b7e6a75ea741de8ec9bf",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}/force_delete": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Force delete user",
                "operationId": "efb44a21077ade6cc278cd50ec689b39",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}